+++ packages/eo-adapters-ogc/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.5](https://github.com/cgi-italy/oida/compare/@oidajs/eo-adapters-ogc@2.4.4...@oidajs/eo-adapters-ogc@2.4.5) (2023-04-13)
+
+**Note:** Version bump only for package @oidajs/eo-adapters-ogc
+
+## [2.4.4](https://github.com/cgi-italy/oida/compare/@oidajs/eo-adapters-ogc@2.4.3...@oidajs/eo-adapters-ogc@2.4.4) (2023-04-12)
+
+**Note:** Version bump only for package @oidajs/eo-adapters-ogc
+
 ## [2.4.3](https://github.com/cgi-italy/oida/compare/@oidajs/eo-adapters-ogc@2.4.2...@oidajs/eo-adapters-ogc@2.4.3) (2023-04-12)
 
 **Note:** Version bump only for package @oidajs/eo-adapters-ogc